Scrape some words from Merriam-Webster
Just the a and b words for now to make testing more efficient. First the following npm packages were installed for scraping: cheerio, request, request-promise. Then the browserify npm package was installed to generate merriamWebsterBundle.js from merrianWebster.js to allow the corresponding NodeJS code to be bundled for execution in the browser.